Minor cleanups to minitest/parallel_fork/interrupt
In minitest/parallel_fork, remove run_after_parallel_fork_setup_hook
method, and implement parallel_fork_fork_child instead. This makes
it so a hook method is not needed when the interrupt code is not used.

Make the following changes to minitest/parallel_fork/interrupt:

* require minitest at the top so this works even if minitest has not
  yet been loaded.

* Replace ||= with = when setting @parallel_fork_child_pids, since this
  is only called once.

* Send Interrupted message to stderr instead of stdout, and combine the
  messages into one string.

* Drop exit 130.  Doesn't seem necessary, and wasn't tested for. Exiting
  non-zero seems sufficient.

* Avoid Ruby 2.5 syntax, since we still target Ruby 2.2+.

Update CHANGELOG and add documentation to the README for this.